Final da Fase 2 - Importar perguntas

+++ carrot-back/src/main/java/com/fernando/carrotback/domain/entity/Game.java
@@ -22,5 +22,5 @@ public class Game {
     @Enumerated(EnumType.STRING)
     private GameStatus status;
     private Long actualQuestion;
-    private Long totalQuestions;
+    private Integer totalQuestions;
 }
